Created /Profile folder and moved profile.js inside
