+++ src/components/Pages/Profile/Profile.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import dva, { connect } from "dva";
-import map from "./images/map.png";
-import prof_pic from "./images/profile_pic.png";
+import map from "../images/map.png";
+import prof_pic from "../images/profile_pic.png";
 import {
   addProfilePhoto,
   addBioText,
   getPictures,
   getBase64ImgFromUrl,
   addInspiration
-} from "../../userApi";
+} from "../../../userApi";
 import {
   Card,
   Upload,
